prediction and update only for dt>0.001

+++ src/FusionEKF.h
@@ -37,8 +37,8 @@ private:
 
   // previous timestamp
   long long previous_timestamp_;
-  int noise_ax;
-  int noise_ay;
+  float noise_ax;
+  float noise_ay;
 
   // tool object used to compute Jacobian and RMSE
   Tools tools;
